requirements.txt
jwt==1.7.1
djangorestframework_simplejwt==4.5.0
